refactor: use InputWrapper around TextField
