+++ src/components/TextField/TextField.test.tsx
@@ -10,17 +10,8 @@ import userEvent from '@testing-library/user-event';
 
 import type { TextFieldProps } from './TextField';
 import { TextField } from './TextField';
-import { ReadOnlyVariant, InputVariant } from './utils';
 
 const user = userEvent.setup();
-
-const getClassNames = (expectedClassName: string) => {
-  const otherClassNames = Object.values(InputVariant).filter(
-    (v) => v !== expectedClassName,
-  );
-
-  return { expectedClassName, otherClassNames };
-};
 
 describe('TextField', () => {
   it('should trigger onPaste when pasting into input', () => {
@@ -63,126 +54,6 @@ describe('TextField', () => {
     await user.keyboard('test');
 
     expect(handleChange).toHaveBeenCalledTimes(4);
-  });
-
-  describe('error-icon', () => {
-    it('should not show error-icon when isValid is true', () => {
-      render({ isValid: true });
-      expect(screen.queryByTestId('input-icon-error')).not.toBeInTheDocument();
-    });
-
-    it('should not show error-icon when isValid is true and readOnly is true', () => {
-      render({ isValid: true, readOnly: true });
-      expect(screen.queryByTestId('input-icon-error')).not.toBeInTheDocument();
-    });
-
-    it('should not show error-icon when isValid is true and disabled is true', () => {
-      render({ isValid: true, disabled: true });
-      expect(screen.queryByTestId('input-icon-error')).not.toBeInTheDocument();
-    });
-
-    it('should show error-icon when isValid is false', () => {
-      render({ isValid: false });
-      expect(screen.queryByTestId('input-icon-error')).toBeInTheDocument();
-    });
-  });
-
-  describe('input-variant', () => {
-    it('should render with correct classname when when isValid is false and readOnly or disabled is not specified', () => {
-      render({ isValid: false });
-      const { expectedClassName, otherClassNames } = getClassNames(
-        InputVariant.Error,
-      );
-
-      const textField = screen.getByTestId('TextField');
-
-      expect(
-        textField.classList.contains(`input-wrapper--${expectedClassName}`),
-      ).toBe(true);
-      otherClassNames.forEach((v) => {
-        expect(textField.classList.contains(`input-wrapper--${v}`)).toBe(false);
-      });
-    });
-
-    it('should render with correct classname when isValid is true and readOnly or disabled is not specified', () => {
-      render({ isValid: true });
-      const { expectedClassName, otherClassNames } = getClassNames(
-        InputVariant.Default,
-      );
-
-      const textField = screen.getByTestId('TextField');
-
-      expect(
-        textField.classList.contains(`input-wrapper--${expectedClassName}`),
-      ).toBe(true);
-      otherClassNames.forEach((v) => {
-        expect(textField.classList.contains(`input-wrapper--${v}`)).toBe(false);
-      });
-    });
-
-    it('should render with correct classname when readOnly is true and disabled is not specified', () => {
-      render({ readOnly: true });
-      const { expectedClassName, otherClassNames } = getClassNames(
-        InputVariant.ReadOnlyInfo,
-      );
-
-      const textField = screen.getByTestId('TextField');
-
-      expect(
-        textField.classList.contains(`input-wrapper--${expectedClassName}`),
-      ).toBe(true);
-      otherClassNames.forEach((v) => {
-        expect(textField.classList.contains(`input-wrapper--${v}`)).toBe(false);
-      });
-    });
-
-    it('should render with correct classname when readOnly is <readonly-confirm> and disabled is not specified', () => {
-      render({ readOnly: ReadOnlyVariant.ReadOnlyConfirm });
-      const { expectedClassName, otherClassNames } = getClassNames(
-        InputVariant.ReadOnlyConfirm,
-      );
-
-      const textField = screen.getByTestId('TextField');
-
-      expect(
-        textField.classList.contains(`input-wrapper--${expectedClassName}`),
-      ).toBe(true);
-      otherClassNames.forEach((v) => {
-        expect(textField.classList.contains(`input-wrapper--${v}`)).toBe(false);
-      });
-    });
-
-    it('should render with correct classname when readOnly is <readonly-info> and disabled is not specified', () => {
-      render({ readOnly: ReadOnlyVariant.ReadOnlyInfo });
-      const { expectedClassName, otherClassNames } = getClassNames(
-        InputVariant.ReadOnlyInfo,
-      );
-
-      const textField = screen.getByTestId('TextField');
-
-      expect(
-        textField.classList.contains(`input-wrapper--${expectedClassName}`),
-      ).toBe(true);
-      otherClassNames.forEach((v) => {
-        expect(textField.classList.contains(`input-wrapper--${v}`)).toBe(false);
-      });
-    });
-
-    it('should render with correct classname when disabled is true', () => {
-      render({ disabled: true });
-      const { expectedClassName, otherClassNames } = getClassNames(
-        InputVariant.Disabled,
-      );
-
-      const textField = screen.getByTestId('TextField');
-
-      expect(
-        textField.classList.contains(`input-wrapper--${expectedClassName}`),
-      ).toBe(true);
-      otherClassNames.forEach((v) => {
-        expect(textField.classList.contains(`input-wrapper--${v}`)).toBe(false);
-      });
-    });
   });
 
   describe('number-format-input', () => {
