+++ src/components/TextField/index.ts
@@ -1,2 +1,1 @@
-export { IconVariant, ReadOnlyVariant } from './utils';
 export { TextField } from './TextField';
